+++ packages/routes/index.js
@@ -5,6 +5,8 @@ import Loadable from 'react-loadable';
 import componentGateway from '@qasir/component-gateway';
 import ListRoute from './list';
 
+import imgNotFound from '../assets/img/not-found.svg';
+
 const loadable = loader => Loadable({
     loader,
     delay: false,
@@ -29,7 +31,16 @@ class Routes extends React.Component {
                     })}
                     <Route
                         render={() => (
-                            <h1>Gak ada Brother</h1>
+                            <div style={{
+                                backgroundColor: '#282c34',
+                                minHeight: '100vh',
+                                display: 'flex',
+                                flexDirection: 'column',
+                                alignItems: 'center',
+                                justifyContent: 'center'
+                            }}>
+                                <img src={imgNotFound} alt="not found" style={{ width: '32%' }}/>
+                            </div>
                         )}
                     />
             </Switch>